Refactor NoOp classes
Attempt to solve Caused by: java.lang.IncompatibleClassChangeError:
Found interface
com.forgerock.openbanking.analytics.services.SessionCountersKPIService,
but class was expected issue
