+++ forgerock-openbanking-analytics-client/src/main/java/com/forgerock/openbanking/analytics/services/NoOpConfiguration.java
@@ -1,0 +1,78 @@
+/**
+ * Copyright 2019 ForgeRock AS.
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.forgerock.openbanking.analytics.services;
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class NoOpConfiguration {
+
+    @ConditionalOnMissingBean
+    @Bean
+    CallBackCountersKPIService noOpCallBackCountersKPIService() {
+        return new NoOpCallBackCountersKPIService();
+    }
+
+    @ConditionalOnMissingBean
+    @Bean
+    ConsentMetricService noOpConsentMetricService() {
+        return new NoOpConsentMetricService();
+    }
+
+    @ConditionalOnMissingBean
+    @Bean
+    DirectoryCountersKPIService noOpDirectoryCountersKPIService() {
+        return new NoOpDirectoryCountersKPIService();
+    }
+
+    @ConditionalOnMissingBean
+    @Bean
+    MetricService noOpMetricService() {
+        return new NoOpMetricService();
+    }
+
+    @ConditionalOnMissingBean
+    @Bean
+    PsuCounterEntryKPIService noOpPsuCounterEntryKPIService() {
+        return new NoOpPsuCounterEntryKPIService();
+    }
+
+    @ConditionalOnMissingBean
+    @Bean
+    SessionCountersKPIService noOpSessionCountersKPIService() {
+        return new NoOpSessionCountersKPIService();
+    }
+
+    @ConditionalOnMissingBean
+    @Bean
+    TokenUsageService noOpTokenUsageService() {
+        return new NoOpTokenUsageService();
+    }
+
+    @ConditionalOnMissingBean
+    @Bean
+    TppEntriesKPIService noOpTppEntriesKPIService() {
+        return new NoOpTppEntriesKPIService();
+    }
+
+}
